Rename market data service to market data provider
