+++ .env.example
@@ -28,4 +28,4 @@ SESSION_LIFETIME=2628000
 
 EXPLORER_URI_PREFIX=payvo
 
-EXPLORER_CRYPTO_MARKET_API_SERVICE="\\App\\Services\\MarketDataServices\\CoinGecko"
+EXPLORER_MARKET_DATA_PROVIDER_SERVICE="\\App\\Services\\MarketDataProviders\\CoinGecko"
