+++ tests/Unit/Services/MarketDataProviders/CryptoCompareTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use App\Services\MarketDataServices\CryptoCompare;
+use App\Services\MarketDataProviders\CryptoCompare;
 use Illuminate\Support\Facades\Http;
 
 use function Spatie\Snapshots\assertMatchesSnapshot;
